Revert "Fixed compile error"
This reverts commit d34aab05279deee3ba645d7f8ba4d3d68b3292f2.

+++ components/taixia/number/taixia_number.h
@@ -15,7 +15,7 @@ class TaiXiaNumber : public number::Number, public TaiXiaListener, public Compon
   void dump_config() override;
 
  protected:
-  void control(double value) override;
+  void control(float value) override;
   uint8_t service_id_;
   uint8_t sa_id_;
